% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_moment_calc.R
\name{tidy_calc_moment}
\alias{tidy_calc_moment}
\title{Calculate community weighted mean, variance, skewness, and kurtosis}
\usage{
tidy_calc_moment(df, trait_names, comm_names, trait_value, weight)
}
\arguments{
\item{df}{A data frame containing the trait and community data.}

\item{trait_names}{Trait id column}

\item{comm_names}{Community names column}

\item{trait_value}{Trait values column}

\item{weight}{Column variable used to weight the trait values}
}
\value{
Returns a data frame with the columns comm, trait, cwm, cwv, cws, and cwk.
}
\description{
Calculates the community weighted mean, variance, skewness, and kurtosis of a given trait.
}
\examples{
df <- data.frame(trait = c("height", "height", "weight", "weight"),
                 trait_value = c(5, 10, 15, 12),
                 abundancia = c(1, 2, 1, 3),
                 comm = c("A", "A", "B", "B"))
tidy_calc_moment(df,
trait_names = trait,
comm_names = comm,
trait_value = trait_value,
weight = abundancia)
}
