% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmoment.R
\docType{package}
\name{ttmoment-package}
\alias{ttmoment}
\alias{ttmoment-package}
\title{ttmoment: Functional Trait Moment Calculation}
\description{
Calculates the community four ‘moments’(mean, variance, skewness, and kurtosis), useful in characterizing the distribution of traits in a community.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/ttmoment}
  \item \url{https://paulesantos.github.io/ttmoment/}
  \item Report bugs at \url{https://github.com/PaulESantos/ttmoment/issues}
}

}
\author{
\strong{Maintainer}: Paul Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

}
\keyword{internal}
